Cellar Views and timestamp

+++ app/Http/Controllers/CellarController.php
@@ -15,9 +15,13 @@ class CellarController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $cellars = Cellar::all();
-        return Inertia::render('CellarList', compact('cellars'));
+    {        
+        //$userId = Auth::id();
+        $userId = 1;
+
+        $cellars = Cellar::where('user_id', $userId)->get();
+
+        return Inertia::render('Cellar/CellarsView', compact('cellars'));
     }
 
     /**
@@ -27,7 +31,7 @@ class CellarController extends Controller
      */
     public function create()
     {
-        return Inertia::render('CellarCreate');
+        return Inertia::render('Cellar/CreateView');
     }
 
     /**
@@ -42,14 +46,15 @@ class CellarController extends Controller
             'name' => 'min:3|max:100'
         ]);
 
-        $userId = Auth::id();
+        //$userId = Auth::id();
+        $userId = 1;
 
         $newCellar = Cellar::create([
             'name' => $request->name,
             'user_id' => $userId
         ]);
 
-        return redirect(route('cellar.show', $newCellar->id))->withSuccess('Cellier créée avec succès');
+        return redirect(route('cellar.show', $newCellar->id))->withSuccess('Cellar successfully created');
     }
 
     /**
@@ -60,13 +65,14 @@ class CellarController extends Controller
      */
     public function show(Cellar $cellar)
     {
-        $userId = Auth::id();
+        //$userId = Auth::id();
+        $userId = 1;
 
         if ($cellar->user_id != $userId) {
-            return redirect(route('cellar.index'))->withErrors("Vous n'avez pas l'autorisation d'accéder à cet cellier");
+            return redirect(route('cellar.index'))->withErrors("You do not have authorization to access this cellar");
         }
 
-        return Inertia::render('CellarShow', compact('cellar'));
+        return Inertia::render('Cellar/ShowView', compact('cellar'));
     }
 
     /**
@@ -77,13 +83,14 @@ class CellarController extends Controller
      */
     public function edit(Cellar $cellar)
     {
-        $userId = Auth::id();
+        //$userId = Auth::id();
+        $userId = 1;
 
         if ($cellar->user_id != $userId) {
-            return redirect(route('cellar.index'))->withErrors("Vous n'avez pas l'autorisation d'accéder à cet cellier");
+            return redirect(route('cellar.index'))->withErrors("You do not have authorization to access this cellar");
         }
 
-        return Inertia::render('CellarEdit', compact('cellar'));
+        return Inertia::render('Cellar/EditView', compact('cellar'));
     }
 
     /**
@@ -105,7 +112,7 @@ class CellarController extends Controller
             'name' => $request->name
         ]);
 
-        return redirect(route('cellar.show', $cellar->id))->withSuccess('Cellier modifiée avec succès');
+        return redirect(route('cellar.show', $cellar->id))->withSuccess('Cellar modified successfully');
     }
 
     /**
@@ -116,9 +123,14 @@ class CellarController extends Controller
      */
     public function destroy(Cellar $cellar)
     {
-        //TO DO: Vérifie s'il y a un enregistrement dans la table Cellar_has_wines avant de le supprimer.
+        $hasWines = $cellar->cellarHasWines()->exists();
+
+        if ($hasWines) {
+            return redirect(route('cellar.index'))->withError('Unable to delete cellar. Wines are associated with it.');
+        }
+
         $cellar->delete();
 
-        return redirect(route('cellar.index'))->withSuccess(trans('Cellier supprimé avec succès'));
+        return redirect(route('cellar.index'))->withSuccess('Cellar deleted successfully');
     }
 }
